Burger Builder Pagination [v1.0.33]
React-router added to the project! Additional pagination has been added, also some extra axios.post() and .get().

Now, the WebApp counts with a full process of a order, well, part of it (there is some dummy–hardcode info, witch is the customer info) and, the orders tab, we can appreciate the orders that has been made by the customer.